notice
# Description

In order to thematically maintain the senpai narrative of son, I
propose keeping in character by changing the implementation of
`imports` to `notice.` Since the current preferred method is importing
`magic` which essentially does the same thing as `imports`, changing
this method to `senpai` will keep with the theme and allow `jsonsenpai`
to notice me.

`imp` is a depreciated lib, so I changed that to `types` instead. The
same functionality takes place. I also cleaned up some formatting
because senpai likes his code neat.

## Before

    with jsonsempai.imports():
        import me

## After

    with jsonsempai.notice():
        import me