Added yaml support
Added YAML support to sempai with the help of the PyYaml parser ; as it is not available in the standard library, all import request will be treated first as a JSON import and only try looking for YAML files if no JSON is to be found. It can't differentiate two files named the same way, so the JSON one will be loaded first.
Changed variable "json_path" to "markup_path" for less confusion.